etl/requirements.txt
farm-haystack[preprocessing,file-conversion,pdf]
numpy
pandas
boto3
python-dotenv
nltk
spacy
beautifulsoup4
langchain==0.0.250
sentence-transformers==2.2.2
chromadb
